First stage of supporting large chromosomes.
The in-memory data structures are 64-bit for pos, mate-pos and insert
size, along with iterators.  The code that fills these out is still
all 32-bit so this is basically a place-holder for ABI purposes.

The exception to this is SAM support, which being purely textual has
the minimal changes necessary to read and write 64-bit values.

Split the hts_parse_reg API to 32-bit and 64-bit variants (although
64 bit version is only used internally at the moment).  To much
code uses this with addresses of 32-bit quantities, so for
compatibility hts_parse_reg() cannot change.

64 bit parse_reg uses a slightly tweaked value for the end for
chromosomes with no range (eg "chr1").  Using INT64_MAX would
yield -1 when cast into int.  We now have nearly 64-bit max which
when truncated to 32-bit is still INT_MAX.

The only change needed in samtools to pass tests is fixing cur5
and pre5 in bam_mate.c.

+++ cram/cram_structs.h
@@ -661,8 +661,8 @@ typedef struct cram_index {
 
 typedef struct {
     int refid;
-    int start;
-    int end;
+    int64_t start;
+    int64_t end;
 } cram_range;
 
 /*-----------------------------------------------------------------------------
